Cleaned up draw function. Started on collisions.
The draw function now mostly consists of recalling other functions to create the game. Just started to do the collision mechanism for platforms. Currently has bugs with the player stopping in mid air when its not only on top of the platform but also while in(colliding with) the platform